backend/python/requirements.txt
fastapi 
uvicorn 
python-multipart 
openai-whisper 
Levenshtein 
ffmpeg-python
tensorflow
numpy
pillow
flask
pandas
google-generativeai